pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "perpetual"
version = "0.1.0"
description = "Autonomous research agent — persistent memory, experiment tracking, hypothesis-driven proposals"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "pygit2>=1.12",
    "lark>=1.1",
    "pyyaml>=6.0",
    "tabulate>=0.9",
]

[project.scripts]
perpetual = "perpetual.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/perpetual"]
